shared geo functions
