+++ src/editor/components/modals/GeoModal/GeoModal.component.jsx
@@ -1,11 +1,8 @@
 import { useState, useCallback, useEffect } from 'react';
 import { SavingModal } from '../SavingModal';
-
 import styles from './GeoModal.module.scss';
 import { Mangnifier20Icon, Save24Icon } from '../../../icons';
-
-import { httpsCallable } from 'firebase/functions';
-import { firebaseConfig, functions } from '../../../services/firebase.js';
+import { firebaseConfig } from '../../../services/firebase.js';
 import Modal from '../Modal.jsx';
 import { Button, Input } from '../../components/index.js';
 import {
@@ -16,6 +13,7 @@ import {
 } from '@react-google-maps/api';
 import GeoImg from '../../../../../ui_assets/geo.png';
 import { roundCoord } from '../../../../../src/utils.js';
+import { setSceneLocation } from '../../../lib/utils.js';
 import useStore from '@/store.js';
 
 const GeoModal = () => {
@@ -54,17 +52,6 @@ const GeoModal = () => {
       }
     }
   }, [isOpen]);
-
-  const requestAndSetElevation = async (lat, lng) => {
-    try {
-      const getGeoidHeight = httpsCallable(functions, 'getGeoidHeight');
-      const result = await getGeoidHeight({ lat: lat, lon: lng });
-      return result.data;
-    } catch (error) {
-      console.error(error.code, error.message, error.details);
-      return null;
-    }
-  };
 
   const setMarkerPositionAndElevation = useCallback((lat, lng) => {
     if (!isNaN(lat) && !isNaN(lng)) {
@@ -117,30 +104,20 @@ const GeoModal = () => {
     setIsWorking(true);
     const latitude = markerPosition.lat;
     const longitude = markerPosition.lng;
-    const data = await requestAndSetElevation(latitude, longitude);
 
-    if (data) {
-      console.log(`latitude: ${latitude}, longitude: ${longitude}`);
-      console.log(`elevation: ${data.ellipsoidalHeight}`);
+    // Use the shared utility function to set the scene location
+    const result = await setSceneLocation(latitude, longitude);
 
-      const geoLayer = document.getElementById('reference-layers');
-      AFRAME.INSPECTOR.execute(
-        geoLayer.hasAttribute('street-geo') ? 'entityupdate' : 'componentadd',
-        {
-          entity: geoLayer,
-          component: 'street-geo',
-          value: {
-            latitude: latitude,
-            longitude: longitude,
-            ellipsoidalHeight: data.ellipsoidalHeight,
-            orthometricHeight: data.orthometricHeight,
-            geoidHeight: data.geoidHeight
-          }
-        }
-      );
-      setTimeout(() => {
-        AFRAME.INSPECTOR.selectEntity(geoLayer);
-      }, 0);
+    if (result.success && result.data) {
+      const data = result.data;
+
+      // Log the new location information
+      console.log('Location data:', data.location);
+      console.log('Location source:', data.locationSource);
+
+      // Log the new intersection information
+      console.log('Nearest intersection:', data.nearestIntersection);
+      console.log('Intersection source:', data.nearestIntersectionSource);
     }
 
     setIsWorking(false);
