+++ src/editor/components/scenegraph/AIChatTools.js
@@ -984,56 +984,19 @@ const AIChatTools = {
   setLatLon: async (args) => {
     const { latitude, longitude } = args;
 
-    if (isNaN(latitude) || isNaN(longitude)) {
-      return 'Error: Invalid latitude or longitude values';
-    }
-
     try {
-      // Import the httpsCallable and functions from firebase
-      const { httpsCallable } = await import('firebase/functions');
-      const { functions } = await import('../../services/firebase.js');
-      const { roundCoord } = await import('../../../../src/utils.js');
+      // Import the setSceneLocation utility function
+      const { setSceneLocation } = await import('../../lib/utils.js');
 
-      // Round coordinates to reasonable precision
-      const lat = roundCoord(parseFloat(latitude));
-      const lng = roundCoord(parseFloat(longitude));
+      // Use the shared utility function to set the scene location
+      const result = await setSceneLocation(latitude, longitude);
 
-      // Request elevation data from the cloud function
-      const getGeoidHeight = httpsCallable(functions, 'getGeoidHeight');
-      const result = await getGeoidHeight({ lat, lon: lng });
-      const data = result.data;
-
-      if (data) {
-        console.log(`Setting location - latitude: ${lat}, longitude: ${lng}`);
-        console.log(`Elevation data: ${JSON.stringify(data)}`);
-
-        // Get the reference layers element
-        const geoLayer = document.getElementById('reference-layers');
-
-        // Update or add the street-geo component
-        AFRAME.INSPECTOR.execute(
-          geoLayer.hasAttribute('street-geo') ? 'entityupdate' : 'componentadd',
-          {
-            entity: geoLayer,
-            component: 'street-geo',
-            value: {
-              latitude: lat,
-              longitude: lng,
-              ellipsoidalHeight: data.ellipsoidalHeight,
-              orthometricHeight: data.orthometricHeight,
-              geoidHeight: data.geoidHeight
-            }
-          }
-        );
-
-        // Select the geo layer in the inspector
-        setTimeout(() => {
-          AFRAME.INSPECTOR.selectEntity(geoLayer);
-        }, 0);
-
-        return `Successfully set location to latitude: ${lat}, longitude: ${lng} with elevation data: ellipsoidal height ${data.ellipsoidalHeight}m, orthometric height ${data.orthometricHeight}m`;
+      // Return appropriate message based on the result
+      if (result.success) {
+        const data = result.data;
+        return `Successfully set location to latitude: ${data.latitude}, longitude: ${data.longitude} with elevation data: ellipsoidal height ${data.ellipsoidalHeight}m, orthometric height ${data.orthometricHeight}m`;
       } else {
-        return 'Error: Failed to retrieve elevation data';
+        return result.message;
       }
     } catch (error) {
       console.error('Error setting lat/lon:', error);
